TVB-2489. Extract dao methods regarding BurstConfiguration2 to a specialized service. Correct fk_parent_burst from DataType to use table BurstConfiguration2. Adjust dependencies towards SimulatorIndex.
